+++ lib/printer_module.dart
@@ -3,3 +3,4 @@ export 'package:blue_thermal_printer/blue_thermal_printer.dart';
 export 'src/print_command.dart';
 export 'src/printer_module.dart';
 export 'src/printer_module_network.dart';
+export 'src/usb_ready_result.dart';